requirements.txt
torch
torchvision
timm
transformers
pandas
pillow
seaborn
scikit-learn
matplotlib